requirements.txt
# Core dependencies
websockets==15.0.1
websocket-client==1.8.0
flask==3.1.2
requests==2.32.5
openai==1.107.3

# Utilities
python-dotenv==1.1.1
# aiohttp==3.9.0
# aiofiles==23.2.1

# Testing
# pytest==7.4   .3
# pytest-asyncio==0.21.1